+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+Unreleased
+
+* add n_iter parameter in NILC algorithm, can fast the calculation on `hp.map2alm`
+
 Release 1.0, 30th October 2024:
 
 * needlets internal linear combination